fix: concurrency problem when multiple users exec same exercise
This commit fixes the concurrency problem when multiple users execute the
same exercise.

The root of the problem is the file creation and deletion. Before
sending the files to the Docker context, the application saves
them on the filesystem. After the execution, the application
deletes the created files.

The problem occurs because the files are created and deleted with
the same name/identifier on multiple sessions. The application
crashes when trying to delete the resource that possibly is
previously deleted.

The solution applied on this commit adds a new identifier to each
request. This identifier is used to create a folder responsible
for retaining all files of each request. After the test execution,
the folder created to save the request files is deleted.

The identifier strategy is a UUID version 4. The probability of
collision is one in 2.71 x 10^18.